requirements.txt
matplotlib==3.10.0
numpy==2.2.2
opencv_python==4.10.0.84
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.6.1
torch==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
ultralytics==8.3.47
